Refactored OsCode, KeyEvent and KeyValue out of the platform specific modules and merged them in keys/mod.rs, which also fixes building on windows
