+++ src/keys/windows.rs
@@ -2,8 +2,9 @@
 
 #[cfg(not(feature = "interception_driver"))]
 use crate::oskbd::*;
-use kanata_keyberon::key_code::*;
 use std::convert::TryFrom;
+
+use super::{OsCode, KeyEvent, KeyValue};
 
 #[allow(unused)]
 mod keys {
@@ -203,562 +204,6 @@ mod keys {
     pub const VK_OEM_CLEAR: u32 = 0xFE;
 }
 pub use keys::*;
-
-/// This is a shameless copy of evdev_rs::enums::EV_KEY.
-/// I've added the Copy trait and I'll be able
-/// to added my own Impl(s) to it
-#[allow(non_camel_case_types)]
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum OsCode {
-    KEY_RESERVED,
-    KEY_ESC,
-    KEY_1,
-    KEY_2,
-    KEY_3,
-    KEY_4,
-    KEY_5,
-    KEY_6,
-    KEY_7,
-    KEY_8,
-    KEY_9,
-    KEY_0,
-    KEY_MINUS,
-    KEY_EQUAL,
-    KEY_BACKSPACE,
-    KEY_TAB,
-    KEY_Q,
-    KEY_W,
-    KEY_E,
-    KEY_R,
-    KEY_T,
-    KEY_Y,
-    KEY_U,
-    KEY_I,
-    KEY_O,
-    KEY_P,
-    KEY_LEFTBRACE,
-    KEY_RIGHTBRACE,
-    KEY_ENTER,
-    KEY_LEFTCTRL,
-    KEY_A,
-    KEY_S,
-    KEY_D,
-    KEY_F,
-    KEY_G,
-    KEY_H,
-    KEY_J,
-    KEY_K,
-    KEY_L,
-    KEY_SEMICOLON,
-    KEY_APOSTROPHE,
-    KEY_GRAVE,
-    KEY_LEFTSHIFT,
-    KEY_BACKSLASH,
-    KEY_Z,
-    KEY_X,
-    KEY_C,
-    KEY_V,
-    KEY_B,
-    KEY_N,
-    KEY_M,
-    KEY_COMMA,
-    KEY_DOT,
-    KEY_SLASH,
-    KEY_RIGHTSHIFT,
-    KEY_KPASTERISK,
-    KEY_LEFTALT,
-    KEY_SPACE,
-    KEY_CAPSLOCK,
-    KEY_F1,
-    KEY_F2,
-    KEY_F3,
-    KEY_F4,
-    KEY_F5,
-    KEY_F6,
-    KEY_F7,
-    KEY_F8,
-    KEY_F9,
-    KEY_F10,
-    KEY_NUMLOCK,
-    KEY_SCROLLLOCK,
-    KEY_KP7,
-    KEY_KP8,
-    KEY_KP9,
-    KEY_KPMINUS,
-    KEY_KP4,
-    KEY_KP5,
-    KEY_KP6,
-    KEY_KPPLUS,
-    KEY_KP1,
-    KEY_KP2,
-    KEY_KP3,
-    KEY_KP0,
-    KEY_KPDOT,
-    KEY_ZENKAKUHANKAKU,
-    KEY_102ND,
-    KEY_F11,
-    KEY_F12,
-    KEY_RO,
-    KEY_KATAKANA,
-    KEY_HIRAGANA,
-    KEY_HENKAN,
-    KEY_KATAKANAHIRAGANA,
-    KEY_MUHENKAN,
-    KEY_KPJPCOMMA,
-    KEY_KPENTER,
-    KEY_RIGHTCTRL,
-    KEY_KPSLASH,
-    KEY_SYSRQ,
-    KEY_RIGHTALT,
-    KEY_LINEFEED,
-    KEY_HOME,
-    KEY_UP,
-    KEY_PAGEUP,
-    KEY_LEFT,
-    KEY_RIGHT,
-    KEY_END,
-    KEY_DOWN,
-    KEY_PAGEDOWN,
-    KEY_INSERT,
-    KEY_DELETE,
-    KEY_MACRO,
-    KEY_MUTE,
-    KEY_VOLUMEDOWN,
-    KEY_VOLUMEUP,
-    KEY_POWER,
-    KEY_KPEQUAL,
-    KEY_KPPLUSMINUS,
-    KEY_PAUSE,
-    KEY_SCALE,
-    KEY_KPCOMMA,
-    KEY_HANGEUL,
-    KEY_HANJA,
-    KEY_YEN,
-    KEY_LEFTMETA,
-    KEY_RIGHTMETA,
-    KEY_COMPOSE,
-    KEY_STOP,
-    KEY_AGAIN,
-    KEY_PROPS,
-    KEY_UNDO,
-    KEY_FRONT,
-    KEY_COPY,
-    KEY_OPEN,
-    KEY_PASTE,
-    KEY_FIND,
-    KEY_CUT,
-    KEY_HELP,
-    KEY_MENU,
-    KEY_CALC,
-    KEY_SETUP,
-    KEY_SLEEP,
-    KEY_WAKEUP,
-    KEY_FILE,
-    KEY_SENDFILE,
-    KEY_DELETEFILE,
-    KEY_XFER,
-    KEY_PROG1,
-    KEY_PROG2,
-    KEY_WWW,
-    KEY_MSDOS,
-    KEY_COFFEE,
-    KEY_ROTATE_DISPLAY,
-    KEY_CYCLEWINDOWS,
-    KEY_MAIL,
-    KEY_BOOKMARKS,
-    KEY_COMPUTER,
-    KEY_BACK,
-    KEY_FORWARD,
-    KEY_CLOSECD,
-    KEY_EJECTCD,
-    KEY_EJECTCLOSECD,
-    KEY_NEXTSONG,
-    KEY_PLAYPAUSE,
-    KEY_PREVIOUSSONG,
-    KEY_STOPCD,
-    KEY_RECORD,
-    KEY_REWIND,
-    KEY_PHONE,
-    KEY_ISO,
-    KEY_CONFIG,
-    KEY_HOMEPAGE,
-    KEY_REFRESH,
-    KEY_EXIT,
-    KEY_MOVE,
-    KEY_EDIT,
-    KEY_SCROLLUP,
-    KEY_SCROLLDOWN,
-    KEY_KPLEFTPAREN,
-    KEY_KPRIGHTPAREN,
-    KEY_NEW,
-    KEY_REDO,
-    KEY_F13,
-    KEY_F14,
-    KEY_F15,
-    KEY_F16,
-    KEY_F17,
-    KEY_F18,
-    KEY_F19,
-    KEY_F20,
-    KEY_F21,
-    KEY_F22,
-    KEY_F23,
-    KEY_F24,
-    KEY_PLAYCD,
-    KEY_PAUSECD,
-    KEY_PROG3,
-    KEY_PROG4,
-    KEY_DASHBOARD,
-    KEY_SUSPEND,
-    KEY_CLOSE,
-    KEY_PLAY,
-    KEY_FASTFORWARD,
-    KEY_BASSBOOST,
-    KEY_PRINT,
-    KEY_HP,
-    KEY_CAMERA,
-    KEY_SOUND,
-    KEY_QUESTION,
-    KEY_EMAIL,
-    KEY_CHAT,
-    KEY_SEARCH,
-    KEY_CONNECT,
-    KEY_FINANCE,
-    KEY_SPORT,
-    KEY_SHOP,
-    KEY_ALTERASE,
-    KEY_CANCEL,
-    KEY_BRIGHTNESSDOWN,
-    KEY_BRIGHTNESSUP,
-    KEY_MEDIA,
-    KEY_SWITCHVIDEOMODE,
-    KEY_KBDILLUMTOGGLE,
-    KEY_KBDILLUMDOWN,
-    KEY_KBDILLUMUP,
-    KEY_SEND,
-    KEY_REPLY,
-    KEY_FORWARDMAIL,
-    KEY_SAVE,
-    KEY_DOCUMENTS,
-    KEY_BATTERY,
-    KEY_BLUETOOTH,
-    KEY_WLAN,
-    KEY_UWB,
-    KEY_UNKNOWN,
-    KEY_VIDEO_NEXT,
-    KEY_VIDEO_PREV,
-    KEY_BRIGHTNESS_CYCLE,
-    KEY_BRIGHTNESS_AUTO,
-    KEY_DISPLAY_OFF,
-    KEY_WWAN,
-    KEY_RFKILL,
-    KEY_MICMUTE,
-    KEY_OK,
-    KEY_SELECT,
-    KEY_GOTO,
-    KEY_CLEAR,
-    KEY_POWER2,
-    KEY_OPTION,
-    KEY_INFO,
-    KEY_TIME,
-    KEY_VENDOR,
-    KEY_ARCHIVE,
-    KEY_PROGRAM,
-    KEY_CHANNEL,
-    KEY_FAVORITES,
-    KEY_EPG,
-    KEY_PVR,
-    KEY_MHP,
-    KEY_LANGUAGE,
-    KEY_TITLE,
-    KEY_SUBTITLE,
-    KEY_ANGLE,
-    KEY_FULL_SCREEN,
-    KEY_MODE,
-    KEY_KEYBOARD,
-    KEY_ASPECT_RATIO,
-    KEY_PC,
-    KEY_TV,
-    KEY_TV2,
-    KEY_VCR,
-    KEY_VCR2,
-    KEY_SAT,
-    KEY_SAT2,
-    KEY_CD,
-    KEY_TAPE,
-    KEY_RADIO,
-    KEY_TUNER,
-    KEY_PLAYER,
-    KEY_TEXT,
-    KEY_DVD,
-    KEY_AUX,
-    KEY_MP3,
-    KEY_AUDIO,
-    KEY_VIDEO,
-    KEY_DIRECTORY,
-    KEY_LIST,
-    KEY_MEMO,
-    KEY_CALENDAR,
-    KEY_RED,
-    KEY_GREEN,
-    KEY_YELLOW,
-    KEY_BLUE,
-    KEY_CHANNELUP,
-    KEY_CHANNELDOWN,
-    KEY_FIRST,
-    KEY_LAST,
-    KEY_AB,
-    KEY_NEXT,
-    KEY_RESTART,
-    KEY_SLOW,
-    KEY_SHUFFLE,
-    KEY_BREAK,
-    KEY_PREVIOUS,
-    KEY_DIGITS,
-    KEY_TEEN,
-    KEY_TWEN,
-    KEY_VIDEOPHONE,
-    KEY_GAMES,
-    KEY_ZOOMIN,
-    KEY_ZOOMOUT,
-    KEY_ZOOMRESET,
-    KEY_WORDPROCESSOR,
-    KEY_EDITOR,
-    KEY_SPREADSHEET,
-    KEY_GRAPHICSEDITOR,
-    KEY_PRESENTATION,
-    KEY_DATABASE,
-    KEY_NEWS,
-    KEY_VOICEMAIL,
-    KEY_ADDRESSBOOK,
-    KEY_MESSENGER,
-    KEY_DISPLAYTOGGLE,
-    KEY_SPELLCHECK,
-    KEY_LOGOFF,
-    KEY_DOLLAR,
-    KEY_EURO,
-    KEY_FRAMEBACK,
-    KEY_FRAMEFORWARD,
-    KEY_CONTEXT_MENU,
-    KEY_MEDIA_REPEAT,
-    KEY_10CHANNELSUP,
-    KEY_10CHANNELSDOWN,
-    KEY_IMAGES,
-    KEY_DEL_EOL,
-    KEY_DEL_EOS,
-    KEY_INS_LINE,
-    KEY_DEL_LINE,
-    KEY_FN,
-    KEY_FN_ESC,
-    KEY_FN_F1,
-    KEY_FN_F2,
-    KEY_FN_F3,
-    KEY_FN_F4,
-    KEY_FN_F5,
-    KEY_FN_F6,
-    KEY_FN_F7,
-    KEY_FN_F8,
-    KEY_FN_F9,
-    KEY_FN_F10,
-    KEY_FN_F11,
-    KEY_FN_F12,
-    KEY_FN_1,
-    KEY_FN_2,
-    KEY_FN_D,
-    KEY_FN_E,
-    KEY_FN_F,
-    KEY_FN_S,
-    KEY_FN_B,
-    KEY_BRL_DOT1,
-    KEY_BRL_DOT2,
-    KEY_BRL_DOT3,
-    KEY_BRL_DOT4,
-    KEY_BRL_DOT5,
-    KEY_BRL_DOT6,
-    KEY_BRL_DOT7,
-    KEY_BRL_DOT8,
-    KEY_BRL_DOT9,
-    KEY_BRL_DOT10,
-    KEY_NUMERIC_0,
-    KEY_NUMERIC_1,
-    KEY_NUMERIC_2,
-    KEY_NUMERIC_3,
-    KEY_NUMERIC_4,
-    KEY_NUMERIC_5,
-    KEY_NUMERIC_6,
-    KEY_NUMERIC_7,
-    KEY_NUMERIC_8,
-    KEY_NUMERIC_9,
-    KEY_NUMERIC_STAR,
-    KEY_NUMERIC_POUND,
-    KEY_NUMERIC_A,
-    KEY_NUMERIC_B,
-    KEY_NUMERIC_C,
-    KEY_NUMERIC_D,
-    KEY_CAMERA_FOCUS,
-    KEY_WPS_BUTTON,
-    KEY_TOUCHPAD_TOGGLE,
-    KEY_TOUCHPAD_ON,
-    KEY_TOUCHPAD_OFF,
-    KEY_CAMERA_ZOOMIN,
-    KEY_CAMERA_ZOOMOUT,
-    KEY_CAMERA_UP,
-    KEY_CAMERA_DOWN,
-    KEY_CAMERA_LEFT,
-    KEY_CAMERA_RIGHT,
-    KEY_ATTENDANT_ON,
-    KEY_ATTENDANT_OFF,
-    KEY_ATTENDANT_TOGGLE,
-    KEY_LIGHTS_TOGGLE,
-    KEY_ALS_TOGGLE,
-    KEY_ROTATE_LOCK_TOGGLE,
-    KEY_BUTTONCONFIG,
-    KEY_TASKMANAGER,
-    KEY_JOURNAL,
-    KEY_CONTROLPANEL,
-    KEY_APPSELECT,
-    KEY_SCREENSAVER,
-    KEY_VOICECOMMAND,
-    KEY_ASSISTANT,
-    KEY_KBD_LAYOUT_NEXT,
-    KEY_BRIGHTNESS_MIN,
-    KEY_BRIGHTNESS_MAX,
-    KEY_KBDINPUTASSIST_PREV,
-    KEY_KBDINPUTASSIST_NEXT,
-    KEY_KBDINPUTASSIST_PREVGROUP,
-    KEY_KBDINPUTASSIST_NEXTGROUP,
-    KEY_KBDINPUTASSIST_ACCEPT,
-    KEY_KBDINPUTASSIST_CANCEL,
-    KEY_RIGHT_UP,
-    KEY_RIGHT_DOWN,
-    KEY_LEFT_UP,
-    KEY_LEFT_DOWN,
-    KEY_ROOT_MENU,
-    KEY_MEDIA_TOP_MENU,
-    KEY_NUMERIC_11,
-    KEY_NUMERIC_12,
-    KEY_AUDIO_DESC,
-    KEY_3D_MODE,
-    KEY_NEXT_FAVORITE,
-    KEY_STOP_RECORD,
-    KEY_PAUSE_RECORD,
-    KEY_VOD,
-    KEY_UNMUTE,
-    KEY_FASTREVERSE,
-    KEY_SLOWREVERSE,
-    KEY_DATA,
-    KEY_ONSCREEN_KEYBOARD,
-    KEY_MAX,
-    BTN_0,
-    BTN_1,
-    BTN_2,
-    BTN_3,
-    BTN_4,
-    BTN_5,
-    BTN_6,
-    BTN_7,
-    BTN_8,
-    BTN_9,
-    BTN_LEFT,
-    BTN_RIGHT,
-    BTN_MIDDLE,
-    BTN_SIDE,
-    BTN_EXTRA,
-    BTN_FORWARD,
-    BTN_BACK,
-    BTN_TASK,
-    BTN_TRIGGER,
-    BTN_THUMB,
-    BTN_THUMB2,
-    BTN_TOP,
-    BTN_TOP2,
-    BTN_PINKIE,
-    BTN_BASE,
-    BTN_BASE2,
-    BTN_BASE3,
-    BTN_BASE4,
-    BTN_BASE5,
-    BTN_BASE6,
-    BTN_DEAD,
-    BTN_SOUTH,
-    BTN_EAST,
-    BTN_C,
-    BTN_NORTH,
-    BTN_WEST,
-    BTN_Z,
-    BTN_TL,
-    BTN_TR,
-    BTN_TL2,
-    BTN_TR2,
-    BTN_SELECT,
-    BTN_START,
-    BTN_MODE,
-    BTN_THUMBL,
-    BTN_THUMBR,
-    BTN_TOOL_PEN,
-    BTN_TOOL_RUBBER,
-    BTN_TOOL_BRUSH,
-    BTN_TOOL_PENCIL,
-    BTN_TOOL_AIRBRUSH,
-    BTN_TOOL_FINGER,
-    BTN_TOOL_MOUSE,
-    BTN_TOOL_LENS,
-    BTN_TOOL_QUINTTAP,
-    BTN_STYLUS3,
-    BTN_TOUCH,
-    BTN_STYLUS,
-    BTN_STYLUS2,
-    BTN_TOOL_DOUBLETAP,
-    BTN_TOOL_TRIPLETAP,
-    BTN_TOOL_QUADTAP,
-    BTN_GEAR_DOWN,
-    BTN_GEAR_UP,
-    BTN_DPAD_UP,
-    BTN_DPAD_DOWN,
-    BTN_DPAD_LEFT,
-    BTN_DPAD_RIGHT,
-    BTN_TRIGGER_HAPPY1,
-    BTN_TRIGGER_HAPPY2,
-    BTN_TRIGGER_HAPPY3,
-    BTN_TRIGGER_HAPPY4,
-    BTN_TRIGGER_HAPPY5,
-    BTN_TRIGGER_HAPPY6,
-    BTN_TRIGGER_HAPPY7,
-    BTN_TRIGGER_HAPPY8,
-    BTN_TRIGGER_HAPPY9,
-    BTN_TRIGGER_HAPPY10,
-    BTN_TRIGGER_HAPPY11,
-    BTN_TRIGGER_HAPPY12,
-    BTN_TRIGGER_HAPPY13,
-    BTN_TRIGGER_HAPPY14,
-    BTN_TRIGGER_HAPPY15,
-    BTN_TRIGGER_HAPPY16,
-    BTN_TRIGGER_HAPPY17,
-    BTN_TRIGGER_HAPPY18,
-    BTN_TRIGGER_HAPPY19,
-    BTN_TRIGGER_HAPPY20,
-    BTN_TRIGGER_HAPPY21,
-    BTN_TRIGGER_HAPPY22,
-    BTN_TRIGGER_HAPPY23,
-    BTN_TRIGGER_HAPPY24,
-    BTN_TRIGGER_HAPPY25,
-    BTN_TRIGGER_HAPPY26,
-    BTN_TRIGGER_HAPPY27,
-    BTN_TRIGGER_HAPPY28,
-    BTN_TRIGGER_HAPPY29,
-    BTN_TRIGGER_HAPPY30,
-    BTN_TRIGGER_HAPPY31,
-    BTN_TRIGGER_HAPPY32,
-    BTN_TRIGGER_HAPPY33,
-    BTN_TRIGGER_HAPPY34,
-    BTN_TRIGGER_HAPPY35,
-    BTN_TRIGGER_HAPPY36,
-    BTN_TRIGGER_HAPPY37,
-    BTN_TRIGGER_HAPPY38,
-    BTN_TRIGGER_HAPPY39,
-    BTN_TRIGGER_HAPPY40,
-    BTN_MAX,
-}
 
 impl OsCode {
     pub fn from_u32(code: u32) -> Option<Self> {
@@ -1039,419 +484,6 @@ impl OsCode {
             OsCode::KEY_RO => 0xC1,
             _ => 0,
         }
-    }
-}
-
-impl TryFrom<usize> for OsCode {
-    type Error = ();
-    fn try_from(item: usize) -> Result<Self, Self::Error> {
-        match Self::from_u32(item as u32) {
-            Some(kc) => Ok(kc),
-            _ => Err(()),
-        }
-    }
-}
-
-impl From<u32> for OsCode {
-    fn from(item: u32) -> Self {
-        Self::from_u32(item).unwrap_or_else(|| panic!("Invalid KeyCode: {}", item))
-    }
-}
-
-impl From<OsCode> for usize {
-    fn from(item: OsCode) -> Self {
-        item.as_u32() as usize
-    }
-}
-
-impl From<OsCode> for u32 {
-    fn from(item: OsCode) -> Self {
-        item.as_u32()
-    }
-}
-
-impl From<KeyCode> for OsCode {
-    fn from(item: KeyCode) -> Self {
-        match item {
-            KeyCode::Escape => OsCode::KEY_ESC,
-            KeyCode::Kb1 => OsCode::KEY_1,
-            KeyCode::Kb2 => OsCode::KEY_2,
-            KeyCode::Kb3 => OsCode::KEY_3,
-            KeyCode::Kb4 => OsCode::KEY_4,
-            KeyCode::Kb5 => OsCode::KEY_5,
-            KeyCode::Kb6 => OsCode::KEY_6,
-            KeyCode::Kb7 => OsCode::KEY_7,
-            KeyCode::Kb8 => OsCode::KEY_8,
-            KeyCode::Kb9 => OsCode::KEY_9,
-            KeyCode::Kb0 => OsCode::KEY_0,
-            KeyCode::Minus => OsCode::KEY_MINUS,
-            KeyCode::Equal => OsCode::KEY_EQUAL,
-            KeyCode::BSpace => OsCode::KEY_BACKSPACE,
-            KeyCode::Tab => OsCode::KEY_TAB,
-            KeyCode::Q => OsCode::KEY_Q,
-            KeyCode::W => OsCode::KEY_W,
-            KeyCode::E => OsCode::KEY_E,
-            KeyCode::R => OsCode::KEY_R,
-            KeyCode::T => OsCode::KEY_T,
-            KeyCode::Y => OsCode::KEY_Y,
-            KeyCode::U => OsCode::KEY_U,
-            KeyCode::I => OsCode::KEY_I,
-            KeyCode::O => OsCode::KEY_O,
-            KeyCode::P => OsCode::KEY_P,
-            KeyCode::LBracket => OsCode::KEY_LEFTBRACE,
-            KeyCode::RBracket => OsCode::KEY_RIGHTBRACE,
-            KeyCode::Enter => OsCode::KEY_ENTER,
-            KeyCode::LCtrl => OsCode::KEY_LEFTCTRL,
-            KeyCode::A => OsCode::KEY_A,
-            KeyCode::S => OsCode::KEY_S,
-            KeyCode::D => OsCode::KEY_D,
-            KeyCode::F => OsCode::KEY_F,
-            KeyCode::G => OsCode::KEY_G,
-            KeyCode::H => OsCode::KEY_H,
-            KeyCode::J => OsCode::KEY_J,
-            KeyCode::K => OsCode::KEY_K,
-            KeyCode::L => OsCode::KEY_L,
-            KeyCode::SColon => OsCode::KEY_SEMICOLON,
-            KeyCode::Quote => OsCode::KEY_APOSTROPHE,
-            KeyCode::Grave => OsCode::KEY_GRAVE,
-            KeyCode::LShift => OsCode::KEY_LEFTSHIFT,
-            KeyCode::Bslash => OsCode::KEY_BACKSLASH,
-            KeyCode::Z => OsCode::KEY_Z,
-            KeyCode::X => OsCode::KEY_X,
-            KeyCode::C => OsCode::KEY_C,
-            KeyCode::V => OsCode::KEY_V,
-            KeyCode::B => OsCode::KEY_B,
-            KeyCode::N => OsCode::KEY_N,
-            KeyCode::M => OsCode::KEY_M,
-            KeyCode::Comma => OsCode::KEY_COMMA,
-            KeyCode::Dot => OsCode::KEY_DOT,
-            KeyCode::Slash => OsCode::KEY_SLASH,
-            KeyCode::RShift => OsCode::KEY_RIGHTSHIFT,
-            KeyCode::KpAsterisk => OsCode::KEY_KPASTERISK,
-            KeyCode::LAlt => OsCode::KEY_LEFTALT,
-            KeyCode::Space => OsCode::KEY_SPACE,
-            KeyCode::CapsLock => OsCode::KEY_CAPSLOCK,
-            KeyCode::F1 => OsCode::KEY_F1,
-            KeyCode::F2 => OsCode::KEY_F2,
-            KeyCode::F3 => OsCode::KEY_F3,
-            KeyCode::F4 => OsCode::KEY_F4,
-            KeyCode::F5 => OsCode::KEY_F5,
-            KeyCode::F6 => OsCode::KEY_F6,
-            KeyCode::F7 => OsCode::KEY_F7,
-            KeyCode::F8 => OsCode::KEY_F8,
-            KeyCode::F9 => OsCode::KEY_F9,
-            KeyCode::F10 => OsCode::KEY_F10,
-            KeyCode::NumLock => OsCode::KEY_NUMLOCK,
-            KeyCode::ScrollLock => OsCode::KEY_SCROLLLOCK,
-            KeyCode::Kp7 => OsCode::KEY_KP7,
-            KeyCode::Kp8 => OsCode::KEY_KP8,
-            KeyCode::Kp9 => OsCode::KEY_KP9,
-            KeyCode::KpMinus => OsCode::KEY_KPMINUS,
-            KeyCode::Kp4 => OsCode::KEY_KP4,
-            KeyCode::Kp5 => OsCode::KEY_KP5,
-            KeyCode::Kp6 => OsCode::KEY_KP6,
-            KeyCode::KpPlus => OsCode::KEY_KPPLUS,
-            KeyCode::Kp1 => OsCode::KEY_KP1,
-            KeyCode::Kp2 => OsCode::KEY_KP2,
-            KeyCode::Kp3 => OsCode::KEY_KP3,
-            KeyCode::Kp0 => OsCode::KEY_KP0,
-            KeyCode::KpDot => OsCode::KEY_KPDOT,
-            KeyCode::F11 => OsCode::KEY_F11,
-            KeyCode::F12 => OsCode::KEY_F12,
-            KeyCode::KpEnter => OsCode::KEY_KPENTER,
-            KeyCode::RCtrl => OsCode::KEY_RIGHTCTRL,
-            KeyCode::KpSlash => OsCode::KEY_KPSLASH,
-            KeyCode::SysReq => OsCode::KEY_SYSRQ,
-            KeyCode::RAlt => OsCode::KEY_RIGHTALT,
-            KeyCode::Home => OsCode::KEY_HOME,
-            KeyCode::Up => OsCode::KEY_UP,
-            KeyCode::PgUp => OsCode::KEY_PAGEUP,
-            KeyCode::Left => OsCode::KEY_LEFT,
-            KeyCode::Right => OsCode::KEY_RIGHT,
-            KeyCode::End => OsCode::KEY_END,
-            KeyCode::Down => OsCode::KEY_DOWN,
-            KeyCode::PgDown => OsCode::KEY_PAGEDOWN,
-            KeyCode::Insert => OsCode::KEY_INSERT,
-            KeyCode::Delete => OsCode::KEY_DELETE,
-            KeyCode::Mute => OsCode::KEY_MUTE,
-            KeyCode::VolDown => OsCode::KEY_VOLUMEDOWN,
-            KeyCode::VolUp => OsCode::KEY_VOLUMEUP,
-            KeyCode::Power => OsCode::KEY_POWER,
-            KeyCode::KpEqual => OsCode::KEY_KPEQUAL,
-            KeyCode::Pause => OsCode::KEY_PAUSE,
-            KeyCode::KpComma => OsCode::KEY_KPCOMMA,
-            KeyCode::LGui => OsCode::KEY_LEFTMETA,
-            KeyCode::RGui => OsCode::KEY_RIGHTMETA,
-            KeyCode::Stop => OsCode::KEY_STOP,
-            KeyCode::Again => OsCode::KEY_AGAIN,
-            KeyCode::Undo => OsCode::KEY_UNDO,
-            KeyCode::Copy => OsCode::KEY_COPY,
-            KeyCode::Paste => OsCode::KEY_PASTE,
-            KeyCode::Find => OsCode::KEY_FIND,
-            KeyCode::Cut => OsCode::KEY_CUT,
-            KeyCode::Help => OsCode::KEY_HELP,
-            KeyCode::Menu => OsCode::KEY_MENU,
-            KeyCode::MediaCalc => OsCode::KEY_CALC,
-            KeyCode::MediaSleep => OsCode::KEY_SLEEP,
-            KeyCode::MediaWWW => OsCode::KEY_WWW,
-            KeyCode::MediaCoffee => OsCode::KEY_COFFEE,
-            KeyCode::MediaBack => OsCode::KEY_BACK,
-            KeyCode::MediaForward => OsCode::KEY_FORWARD,
-            KeyCode::MediaEjectCD => OsCode::KEY_EJECTCD,
-            KeyCode::MediaNextSong => OsCode::KEY_NEXTSONG,
-            KeyCode::MediaPlayPause => OsCode::KEY_PLAYPAUSE,
-            KeyCode::MediaPreviousSong => OsCode::KEY_PREVIOUSSONG,
-            KeyCode::MediaStopCD => OsCode::KEY_STOPCD,
-            KeyCode::MediaRefresh => OsCode::KEY_REFRESH,
-            KeyCode::MediaEdit => OsCode::KEY_EDIT,
-            KeyCode::MediaScrollUp => OsCode::KEY_SCROLLUP,
-            KeyCode::MediaScrollDown => OsCode::KEY_SCROLLDOWN,
-            KeyCode::F13 => OsCode::KEY_F13,
-            KeyCode::F14 => OsCode::KEY_F14,
-            KeyCode::F15 => OsCode::KEY_F15,
-            KeyCode::F16 => OsCode::KEY_F16,
-            KeyCode::F17 => OsCode::KEY_F17,
-            KeyCode::F18 => OsCode::KEY_F18,
-            KeyCode::F19 => OsCode::KEY_F19,
-            KeyCode::F20 => OsCode::KEY_F20,
-            KeyCode::F21 => OsCode::KEY_F21,
-            KeyCode::F22 => OsCode::KEY_F22,
-            KeyCode::F23 => OsCode::KEY_F23,
-            KeyCode::F24 => OsCode::KEY_F24,
-            KeyCode::Lang1 => OsCode::KEY_HANGEUL,
-            KeyCode::Lang2 => OsCode::KEY_HANJA,
-            KeyCode::NonUsBslash => OsCode::KEY_102ND,
-            KeyCode::PScreen => OsCode::KEY_PRINT,
-            KeyCode::Application => OsCode::KEY_COMPOSE,
-            KeyCode::AltErase => OsCode::KEY_ALTERASE,
-            KeyCode::Cancel => OsCode::KEY_CANCEL,
-            KeyCode::MediaMute => OsCode::KEY_MICMUTE,
-            KeyCode::Intl1 => OsCode::KEY_RO,
-            _ => OsCode::KEY_UNKNOWN,
-        }
-    }
-}
-
-impl From<&KeyCode> for OsCode {
-    fn from(item: &KeyCode) -> Self {
-        (*item).into()
-    }
-}
-
-impl From<OsCode> for KeyCode {
-    fn from(item: OsCode) -> KeyCode {
-        match item {
-            OsCode::KEY_ESC => KeyCode::Escape,
-            OsCode::KEY_1 => KeyCode::Kb1,
-            OsCode::KEY_2 => KeyCode::Kb2,
-            OsCode::KEY_3 => KeyCode::Kb3,
-            OsCode::KEY_4 => KeyCode::Kb4,
-            OsCode::KEY_5 => KeyCode::Kb5,
-            OsCode::KEY_6 => KeyCode::Kb6,
-            OsCode::KEY_7 => KeyCode::Kb7,
-            OsCode::KEY_8 => KeyCode::Kb8,
-            OsCode::KEY_9 => KeyCode::Kb9,
-            OsCode::KEY_0 => KeyCode::Kb0,
-            OsCode::KEY_MINUS => KeyCode::Minus,
-            OsCode::KEY_EQUAL => KeyCode::Equal,
-            OsCode::KEY_BACKSPACE => KeyCode::BSpace,
-            OsCode::KEY_TAB => KeyCode::Tab,
-            OsCode::KEY_Q => KeyCode::Q,
-            OsCode::KEY_W => KeyCode::W,
-            OsCode::KEY_E => KeyCode::E,
-            OsCode::KEY_R => KeyCode::R,
-            OsCode::KEY_T => KeyCode::T,
-            OsCode::KEY_Y => KeyCode::Y,
-            OsCode::KEY_U => KeyCode::U,
-            OsCode::KEY_I => KeyCode::I,
-            OsCode::KEY_O => KeyCode::O,
-            OsCode::KEY_P => KeyCode::P,
-            OsCode::KEY_LEFTBRACE => KeyCode::LBracket,
-            OsCode::KEY_RIGHTBRACE => KeyCode::RBracket,
-            OsCode::KEY_ENTER => KeyCode::Enter,
-            OsCode::KEY_LEFTCTRL => KeyCode::LCtrl,
-            OsCode::KEY_A => KeyCode::A,
-            OsCode::KEY_S => KeyCode::S,
-            OsCode::KEY_D => KeyCode::D,
-            OsCode::KEY_F => KeyCode::F,
-            OsCode::KEY_G => KeyCode::G,
-            OsCode::KEY_H => KeyCode::H,
-            OsCode::KEY_J => KeyCode::J,
-            OsCode::KEY_K => KeyCode::K,
-            OsCode::KEY_L => KeyCode::L,
-            OsCode::KEY_SEMICOLON => KeyCode::SColon,
-            OsCode::KEY_APOSTROPHE => KeyCode::Quote,
-            OsCode::KEY_GRAVE => KeyCode::Grave,
-            OsCode::KEY_LEFTSHIFT => KeyCode::LShift,
-            OsCode::KEY_BACKSLASH => KeyCode::Bslash,
-            OsCode::KEY_Z => KeyCode::Z,
-            OsCode::KEY_X => KeyCode::X,
-            OsCode::KEY_C => KeyCode::C,
-            OsCode::KEY_V => KeyCode::V,
-            OsCode::KEY_B => KeyCode::B,
-            OsCode::KEY_N => KeyCode::N,
-            OsCode::KEY_M => KeyCode::M,
-            OsCode::KEY_COMMA => KeyCode::Comma,
-            OsCode::KEY_DOT => KeyCode::Dot,
-            OsCode::KEY_SLASH => KeyCode::Slash,
-            OsCode::KEY_RIGHTSHIFT => KeyCode::RShift,
-            OsCode::KEY_KPASTERISK => KeyCode::KpAsterisk,
-            OsCode::KEY_LEFTALT => KeyCode::LAlt,
-            OsCode::KEY_SPACE => KeyCode::Space,
-            OsCode::KEY_CAPSLOCK => KeyCode::CapsLock,
-            OsCode::KEY_F1 => KeyCode::F1,
-            OsCode::KEY_F2 => KeyCode::F2,
-            OsCode::KEY_F3 => KeyCode::F3,
-            OsCode::KEY_F4 => KeyCode::F4,
-            OsCode::KEY_F5 => KeyCode::F5,
-            OsCode::KEY_F6 => KeyCode::F6,
-            OsCode::KEY_F7 => KeyCode::F7,
-            OsCode::KEY_F8 => KeyCode::F8,
-            OsCode::KEY_F9 => KeyCode::F9,
-            OsCode::KEY_F10 => KeyCode::F10,
-            OsCode::KEY_NUMLOCK => KeyCode::NumLock,
-            OsCode::KEY_SCROLLLOCK => KeyCode::ScrollLock,
-            OsCode::KEY_KP7 => KeyCode::Kp7,
-            OsCode::KEY_KP8 => KeyCode::Kp8,
-            OsCode::KEY_KP9 => KeyCode::Kp9,
-            OsCode::KEY_KPMINUS => KeyCode::KpMinus,
-            OsCode::KEY_KP4 => KeyCode::Kp4,
-            OsCode::KEY_KP5 => KeyCode::Kp5,
-            OsCode::KEY_KP6 => KeyCode::Kp6,
-            OsCode::KEY_KPPLUS => KeyCode::KpPlus,
-            OsCode::KEY_KP1 => KeyCode::Kp1,
-            OsCode::KEY_KP2 => KeyCode::Kp2,
-            OsCode::KEY_KP3 => KeyCode::Kp3,
-            OsCode::KEY_KP0 => KeyCode::Kp0,
-            OsCode::KEY_KPDOT => KeyCode::KpDot,
-            OsCode::KEY_F11 => KeyCode::F11,
-            OsCode::KEY_F12 => KeyCode::F12,
-            OsCode::KEY_CLEAR => KeyCode::Clear,
-            OsCode::KEY_KPENTER => KeyCode::KpEnter,
-            OsCode::KEY_RIGHTCTRL => KeyCode::RCtrl,
-            OsCode::KEY_KPSLASH => KeyCode::KpSlash,
-            OsCode::KEY_SYSRQ => KeyCode::SysReq,
-            OsCode::KEY_RIGHTALT => KeyCode::RAlt,
-            OsCode::KEY_HOME => KeyCode::Home,
-            OsCode::KEY_UP => KeyCode::Up,
-            OsCode::KEY_PAGEUP => KeyCode::PgUp,
-            OsCode::KEY_LEFT => KeyCode::Left,
-            OsCode::KEY_RIGHT => KeyCode::Right,
-            OsCode::KEY_END => KeyCode::End,
-            OsCode::KEY_DOWN => KeyCode::Down,
-            OsCode::KEY_PAGEDOWN => KeyCode::PgDown,
-            OsCode::KEY_INSERT => KeyCode::Insert,
-            OsCode::KEY_DELETE => KeyCode::Delete,
-            OsCode::KEY_MUTE => KeyCode::Mute,
-            OsCode::KEY_VOLUMEDOWN => KeyCode::VolDown,
-            OsCode::KEY_VOLUMEUP => KeyCode::VolUp,
-            OsCode::KEY_POWER => KeyCode::Power,
-            OsCode::KEY_KPEQUAL => KeyCode::KpEqual,
-            OsCode::KEY_PAUSE => KeyCode::Pause,
-            OsCode::KEY_KPCOMMA => KeyCode::KpComma,
-            OsCode::KEY_LEFTMETA => KeyCode::LGui,
-            OsCode::KEY_RIGHTMETA => KeyCode::RGui,
-            OsCode::KEY_STOP => KeyCode::Stop,
-            OsCode::KEY_AGAIN => KeyCode::Again,
-            OsCode::KEY_UNDO => KeyCode::Undo,
-            OsCode::KEY_COPY => KeyCode::Copy,
-            OsCode::KEY_PASTE => KeyCode::Paste,
-            OsCode::KEY_FIND => KeyCode::Find,
-            OsCode::KEY_CUT => KeyCode::Cut,
-            OsCode::KEY_HELP => KeyCode::Help,
-            OsCode::KEY_MENU => KeyCode::Menu,
-            OsCode::KEY_CALC => KeyCode::MediaCalc,
-            OsCode::KEY_SLEEP => KeyCode::MediaSleep,
-            OsCode::KEY_WWW => KeyCode::MediaWWW,
-            OsCode::KEY_COFFEE => KeyCode::MediaCoffee,
-            OsCode::KEY_BACK => KeyCode::MediaBack,
-            OsCode::KEY_FORWARD => KeyCode::MediaForward,
-            OsCode::KEY_EJECTCD => KeyCode::MediaEjectCD,
-            OsCode::KEY_NEXTSONG => KeyCode::MediaNextSong,
-            OsCode::KEY_PLAYPAUSE => KeyCode::MediaPlayPause,
-            OsCode::KEY_PREVIOUSSONG => KeyCode::MediaPreviousSong,
-            OsCode::KEY_STOPCD => KeyCode::MediaStopCD,
-            OsCode::KEY_REFRESH => KeyCode::MediaRefresh,
-            OsCode::KEY_EDIT => KeyCode::MediaEdit,
-            OsCode::KEY_SCROLLUP => KeyCode::MediaScrollUp,
-            OsCode::KEY_SCROLLDOWN => KeyCode::MediaScrollDown,
-            OsCode::KEY_F13 => KeyCode::F13,
-            OsCode::KEY_F14 => KeyCode::F14,
-            OsCode::KEY_F15 => KeyCode::F15,
-            OsCode::KEY_F16 => KeyCode::F16,
-            OsCode::KEY_F17 => KeyCode::F17,
-            OsCode::KEY_F18 => KeyCode::F18,
-            OsCode::KEY_F19 => KeyCode::F19,
-            OsCode::KEY_F20 => KeyCode::F20,
-            OsCode::KEY_F21 => KeyCode::F21,
-            OsCode::KEY_F22 => KeyCode::F22,
-            OsCode::KEY_F23 => KeyCode::F23,
-            OsCode::KEY_F24 => KeyCode::F24,
-            OsCode::KEY_HANGEUL => KeyCode::Lang1,
-            OsCode::KEY_HANJA => KeyCode::Lang2,
-            OsCode::KEY_102ND => KeyCode::NonUsBslash,
-            OsCode::KEY_PRINT => KeyCode::PScreen,
-            OsCode::KEY_COMPOSE => KeyCode::Application,
-            OsCode::KEY_ALTERASE => KeyCode::AltErase,
-            OsCode::KEY_CANCEL => KeyCode::Cancel,
-            OsCode::KEY_MICMUTE => KeyCode::MediaMute,
-            OsCode::KEY_RO => KeyCode::Intl1,
-            _ => KeyCode::No,
-        }
-    }
-}
-
-impl From<&OsCode> for KeyCode {
-    fn from(item: &OsCode) -> KeyCode {
-        (*item).into()
-    }
-}
-
-// ------------------ KeyValue --------------------
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum KeyValue {
-    Release = 0,
-    Press = 1,
-    Repeat = 2,
-}
-
-impl From<i32> for KeyValue {
-    fn from(item: i32) -> Self {
-        match item {
-            0 => Self::Release,
-            1 => Self::Press,
-            2 => Self::Repeat,
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl From<bool> for KeyValue {
-    fn from(up: bool) -> Self {
-        match up {
-            true => Self::Release,
-            false => Self::Press,
-        }
-    }
-}
-
-impl From<KeyValue> for bool {
-    fn from(val: KeyValue) -> Self {
-        matches!(val, KeyValue::Release)
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct KeyEvent {
-    pub code: OsCode,
-    pub value: KeyValue,
-}
-
-#[cfg(not(feature = "interception_driver"))]
-impl KeyEvent {
-    pub fn new(code: OsCode, value: KeyValue) -> Self {
-        Self { code, value }
     }
 }
